Moved logic from DataRequestService to ResponseBuilder.
