+++ Service/DataRequestService.php
@@ -17,7 +17,6 @@ use ONGR\ElasticsearchBundle\ORM\Manager;
 use ONGR\ElasticsearchBundle\ORM\Repository;
 use Symfony\Component\DependencyInjection\Container;
 use Symfony\Component\HttpFoundation\Request;
-use Symfony\Component\HttpFoundation\Response;
 
 /**
  * Manages requests to a Document repository.
@@ -72,7 +71,7 @@ class DataRequestService
     public function get($request)
     {
         /** @var Search $search */
-        $search = $this->dataRepository->createSearch();
+        $search = $this->getDataRepository()->createSearch();
 
         $query = new MatchAllQuery();
         $search->addQuery($query);
@@ -83,77 +82,6 @@ class DataRequestService
         }
 
         return $this->dataRepository->execute($search, 'array');
-    }
-
-    /**
-     * Creates response for get requests.
-     *
-     * @param Request $request
-     *
-     * @return Response
-     */
-    public function getResponse(Request $request)
-    {
-        $format = $this->getResponseFormat($request);
-        $mime = $this->getResponseMime($format);
-
-        $data = $this->get($request);
-        $data = $this->encodeArray($data, $format);
-
-        $response = new Response();
-        $response->setContent($data);
-        $response->headers->set('Content-Type', $mime);
-
-        return $response;
-    }
-
-    /**
-     * Encodes array to given format.
-     *
-     * @param array  $data
-     * @param string $format
-     *
-     * @return mixed
-     * @throws \DomainException
-     */
-    private function encodeArray($data, $format)
-    {
-        switch($format){
-            case 'json':
-                return json_encode($data);
-            default:
-                throw new \DomainException("Unknown format \"{$format}\"");
-        }
-    }
-
-    /**
-     * @param Request $request
-     *
-     * @return string
-     */
-    private function getResponseFormat(Request $request)
-    {
-        return 'json';
-    }
-
-    /**
-     * @param string $format
-     *
-     * @return string
-     *
-     * @throws \DomainException
-     */
-    private function getResponseMime($format)
-    {
-        static $types = [
-            'json' => 'application/json',
-        ];
-
-        if (!isset($types[$format])) {
-            throw new \DomainException("Unknown format \"{$format}\"");
-        }
-
-        return $types[$format];
     }
 
     /**
